Add CLI flag to create/append to SAR v2.00 archive
Probably not that useful on day to day usage, but tremendously helps
debugging and testing compatibility between Py 2/3 and official tooling.
Flag only affects creation and appending, extraction works regardless of
flag and the archive version. When appending to an existing archive,
convert the archive to the given version if it differs from the old
version.